Explain advantage of (e,s) better
It has been pointed out on bitcoin-dev that it's actually not a huge
problem to encode a point. And yes, it's not a huge problem but just
a minor addition of complexity. But it should be noted because it's
the only advantage of (e,s), and if we remove it, there will be no
trade-off left that needs to be discussed in the Design section.